[chore] Ensure accurate error reporting if beforeAll callback fails
This fixes an issue that reported inaccurate error message if something failed during `beforeAll`: E.g. if launching `qawolf` in the `beforeAll`-callback failed, the `browser` variable would be undefined, which again would make the `afterAll` callback fail when it tried to call `close()` on it. This obscured the original error, and made jest report `TypeError: Cannot read property 'close' of undefined` as the error and not the original error that happened in the beforeAll`-callback.