fix(docs): add `shallow: true` to query changes in arcade (#551)
* Adding shallow: true to query changes in arcade

I was playing around with the Arcade and noticed it fires requests at every keystroke. That's because the usage of Next's router.replace doesn't include a shallow: true option.

If we add that, Next won't try to query the backend every time the router changes, which will improve performance of the query encoding mechanism and save a lot of requests from being made. [More on Next's docs](https://nextjs.org/docs/routing/shallow-routing).

* fix(docs): fix eslint issues

Co-authored-by: Marius Lundgård <studio@mariuslundgard.com>