feat: dialog and layer improvements (#1052)
This pull request adds improvements and fixes to the `Dialog` component
and the `Layer` util.

Updates in `Layer`:
- Adds an `onActivate` callback that is triggered when the layer becomes
the top layer. The callback returns the last focused element since
`Layer` was the top layer. This is to enable refocusing of the last
focused element when it becomes the top layer again.
- Fixes an issue where the `isTopLayer` value from the `useLayer` hook
was calculated incorrectly

Updates in `Dialog`:
- Allows closing a specific dialog when multiple dialogs are open at the
same time
- Forwards the `onActivate` callback to `Layer` to allow refocusing of
the last focused element when a dialog becomes the top layer.

Co-authored-by: Marius Lundgård <studio@mariuslundgard.com>