feat: allow hovered styles to be applied via the `data-hovered` attr (#1073)
This PR enables hover styles for `<Button>`, `<Card>` and _selectable_
components (`<MenuGroup>`, `<MenuItem>`) to be displayed when
`data-hovered` is set. This is similar to how `data-selected` works on
these components.

Also similar to `data-selected`, the presence of `data-disabled` will
override this behaviour, and these styles will also not be displayed if
no hover support is detected.

This is in service of command lists in the studio that want to
ergonomically set hover state on list items when navigating via the
keyboard (e.g. global search).