fix(tooltip): use floatingStyles to prevent rerenders of tooltip (#1119)
The tooltip would rerender unnecessarily when hovering over it.

The `x` positioning used in `rootstyles` was most likely the reason for
this, as the `x` position would be recalculated several times, causing
the tooltip placement to "jump".
By using the `floatingStyles` from `floating-ui', this was fixed. The
size of the tooltip is calculated using the `size` from `floating-ui`.

---------

Co-authored-by: Marius Lundgård <studio@mariuslundgard.com>